ocr_model/src/inference.py
"""
Inference script for OCR model
Performs OCR on new images
"""

import torch
import cv2
import numpy as np
from pathlib import Path
from PIL import Image
import argparse


class OCRInference:
    """OCR inference engine"""
    
    def __init__(self, model_path, model_class, charset, device='cuda'):
        """
        Initialize inference engine
        
        Args:
            model_path: Path to saved model
            model_class: Model class to instantiate
            charset: Character set for decoding
            device: Device to run on
        """
        self.device = torch.device(device if torch.cuda.is_available() else 'cpu')
        
        # Load model
        self.model = model_class()
        checkpoint = torch.load(model_path, map_location=self.device)
        self.model.load_state_dict(checkpoint['model_state_dict'])
        self.model.to(self.device)
        self.model.eval()
        
        self.charset = charset
        self.idx2char = {idx: char for idx, char in enumerate(charset)}
    
    def preprocess_image(self, image_path, height=64, width=256):
        """Preprocess image for inference"""
        image = cv2.imread(str(image_path))
        if image is None:
            raise ValueError(f"Cannot load image: {image_path}")
        
        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        image = cv2.resize(image, (width, height))
        image = image.astype(np.float32) / 255.0
        image = torch.from_numpy(image).permute(2, 0, 1).unsqueeze(0)
        
        return image.to(self.device)
    
    def predict(self, image_path, confidence_threshold=0.5):
        """
        Predict text from image
        
        Args:
            image_path: Path to image
            confidence_threshold: Minimum confidence threshold
            
        Returns:
            Predicted text
        """
        image = self.preprocess_image(image_path)
        
        with torch.no_grad():
            outputs = self.model(image)
            predictions = torch.softmax(outputs, dim=2)
            predicted_indices = torch.argmax(predictions, dim=2)
        
        # Decode predictions
        text = ''
        for idx in predicted_indices[0]:
            text += self.idx2char.get(idx.item(), '')
        
        return text.strip()


def main():
    parser = argparse.ArgumentParser(description='OCR Inference')
    parser.add_argument('--model-path', required=True, help='Path to saved model')
    parser.add_argument('--image-path', required=True, help='Path to image')
    parser.add_argument('--device', choices=['cuda', 'cpu'], default='cuda', help='Device')
    
    args = parser.parse_args()
    
    # This is a template for inference
    print(f"Model path: {args.model_path}")
    print(f"Image path: {args.image_path}")
    print("Inference script template ready!")


if __name__ == '__main__':
    main()
