ocr_model/requirements.txt
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0
paddleocr>=2.7.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0
albumentations>=1.3.0
